+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [semantic versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 ### Added
+- Added `write_csv` function inside `cdp/helpers/s3_utils.py`.
 
 ### Changed